finalize id3 mapping
this first version of the id3 mapping should be able to keep some
of the id3 data. I plan on using the data stred by this mapping to
get some data on how to improve the mapping and also to look for
glaring errors in the data as well as for some time measurements
to ensure that i haven't sacrificed any performance since i started
rewriting to smw.